openai-terminal/requirements.txt
anima-labs>=0.1.0
openai-agents>=0.1.0
python-dotenv>=1.0.0
